create demos folder * add html output support to webpack * move index.ts to demo1.ts
