+++ .gitignore
@@ -63,5 +63,4 @@ typings/
 # Ignore build ts files
 dist
 
-# netlify hosting
-public
+public/dist
