+++ index.html
@@ -1,10 +1,9 @@
 <!DOCTYPE html>
 <html>
-    <head>
-          <canvas id="render-area" width="800" height="600"></canvas>
-          <script src="dist/web-bundle.js" type="text/javascript" async></script>
-          <link rel="stylesheet" href="style.css" />
-    </head>
-    <body>
-    </body>
+  <head> </head>
+  <body>
+    <h1>Demos</h1>
+    <h2>Demo 1</h2>
+    <canvas id="demo-1" height="800" width="600"></canvas>
+  </body>
 </html>
