+++ package.json
@@ -6,7 +6,7 @@
   "scripts": {
     "build": "tsc",
     "build+watch": "tsc -w",
-    "build:netlify": "sh netlify_build.sh",
+    "build:netlify": "webpack --progress",
     "start-web": "webpack-dev-server",
     "postinstall": "tsc"
   },
@@ -28,6 +28,8 @@
   },
   "devDependencies": {
     "@types/uuid": "^8.0.0",
+    "html-loader": "^1.1.0",
+    "html-webpack-plugin": "^4.3.0",
     "ts-loader": "^7.0.5",
     "webpack": "^4.43.0",
     "webpack-cli": "^3.3.11",
